requirements.txt
streamlit
huggingface_hub
python-dotenv
textblob
watchdog
